Memory improvements

+++ headers/structs.h
@@ -15,15 +15,15 @@ static struct{
 
 typedef struct{
     Texture tx;
-    float x,y;
-    float speedX, speedY; //used for gravity and jumping
-    float cooldown;
+    short x,y;
+    short speedX, speedY; //used for gravity and jumping
+    int8_t cooldown;
     bool falling;
 } Soldier;
 
 typedef struct{
     Texture tx;
-    float x,y;
+    short x,y;
     unsigned short rotation;
     short speedX,speedY;
     bool collided;
@@ -31,22 +31,22 @@ typedef struct{
 
 typedef struct{
     Texture tx;
-    float x,y;
+    short x,y;
     unsigned short rotation;
     int8_t flip; //1 means befault, -1 means flipped
 } Launcher;
 
 typedef struct{
     Texture tx;
-    float x,y;
+    short x,y;
     unsigned short rotation;
     u_int8_t alpha;
-    float cooldownAlpha;
+    int8_t cooldownAlpha;
 } Particle;
 
 typedef struct{
     Texture tx;
-    float x,y;
+    short x,y;
 } Platform;
 
 #endif
